+++ src/os2datascanner/projects/report/reportapp/utils.py
@@ -12,6 +12,9 @@ from django.utils.translation import ugettext_lazy as _
 
 from os2datascanner.engine2.pipeline import messages
 from .models.documentreport import DocumentReport
+from .models.roles.role import Role
+from .models.roles.dpo import DataProtectionOfficer
+from .models.roles.leader import Leader
 
 from os2datascanner.projects.report.organizations.models import (
     Alias, AliasType, Organization, Account)
@@ -105,6 +108,14 @@ def user_is(roles, role_cls):
     """Checks whether a list of roles contains a certain role type (role_cls)"""
     return any(isinstance(role, role_cls)
                for role in roles)
+
+
+def user_is_superadmin(user):
+    """Relevant users to notify if matches exist with
+    the `only_notify_superuser`-flag."""
+    roles = Role.get_user_roles_or_default(user)
+    return (user_is(roles, DataProtectionOfficer)
+            or user_is(roles, Leader) or user.is_superuser)
 
 
 class OIDCAuthenticationBackend(auth.OIDCAuthenticationBackend):