+++ src/os2datascanner/projects/static/js/leaderOverview.js
@@ -1,0 +1,27 @@
+function showOverview(row, toggleButton) {
+
+  let overviewRow = row.nextElementSibling;
+  toggleClass(toggleButton, "up");
+  let buttonOpen = hasClass(toggleButton, "up");
+
+  overviewRow.hidden = !buttonOpen;
+}
+
+document.addEventListener("DOMContentLoaded",
+  () => {
+    htmx.onLoad(function (content) {
+      if (hasClass(content, "content") || hasClass(content, "page")) {
+
+        let expandButtons = document.querySelectorAll(".overview-expand");
+
+        expandButtons.forEach(element => {
+          element.addEventListener("click", function (e) {
+            targ = e.target;
+            let row = closestElement(targ, "tr");
+            showOverview(row, targ);
+          });
+        });
+      }
+    });
+  }
+);