+++ src/os2datascanner/projects/static/js/userCharts.js
@@ -1,0 +1,26 @@
+function drawUserChart(content) {
+
+  const matchesByWeek = JSON.parse(content.querySelector('#matches_by_week').textContent);
+  const accountUuid = JSON.parse(content.querySelector('#uuid').textContent);
+
+  let weeknums = [];
+  let matches = [];
+
+  for (let obj of matchesByWeek) {
+    // Reverse the direction of the arrays with unshift.
+    weeknums.unshift(obj.weeknum);
+    matches.unshift(obj.matches);
+  }
+
+  const UserChartCtx = document.querySelector("#line_chart_all_matches_development__" + accountUuid).getContext('2d');
+
+  makeLineChart(weeknums, matches, UserChartCtx);
+
+
+}
+
+htmx.onLoad((content) => {
+  if (hasClass(content, 'overview_wrapper')) {
+    drawUserChart(content);
+  }
+});